Add files via upload
OOPJ Assignment
1. Implement a class for a “Book”. Book contains a title (a String), a list of authors (array of authors), number of
pages (an integer), price (floating point number), publisher (a String) etc. Write suitable constructor and
accessor/modifier methods. Implement a class for “Library”. A library contains a list of books (array of Book).
Write add (to add a book) and remove (to delete a book) methods for library. Write a main() function to create a
“Library” and add five “Book” to library. Print the total price of all books.
2. Implement a class for “DateDemo”. Write member functions for (i) getting the previous day, (iv) getting the next
day, (iii) printing a day There should be four constructors: (i) day, month and year are initialized to 01, 01, 1970; (ii)
day is initialized to user specified value but month and year are initialized to 01, 1970; (iii) day, month are
initialized to user specified value but year is initialized to 1970; (iv) day, month and year are initialized to user
defined values. Also, write a main() function to (i) create a date object; (ii) print the next and the previous day.
3. Write a class to represent complex numbers with necessary constructors. Write member functions to add, multiply
two complex numbers. There should be three constructors: (i) to initialize real and imaginary parts to 0; (ii) to
initialize imaginary part to 0 but to initialize the real part to user defined value; (iii) to initialize the real part and the
imaginary part to user defined values. Also, write a main() function to (i) create two complex numbers 3+2i and 4-
2i; (ii) to print the sum and product of those numbers.
4. Create a class “Room” which will hold the “height”, “width” and “breadth” of the room in three fields. This class
also has a method “volume()” to calculate the volume of this room. Create another class “RoomDemo” which will
use the earlier class, create instances of rooms, and display the volume of room.
5. Write a program to implement a class “student” with the following members. Name of the student. Marks of the
student obtained in three subjects. Function to assign initial values. Function to compute total average. Function to
display the student’s name and the total marks. Write an appropriate main() function to demonstrate the functioning
of the above.
6. Write a class “Box” that with three member-variables “height”, “width” and “breadth”. Write suitable
constructors to initialize them. Add functions like “getVolume” and “getArea” that will return volume and surface
area respectively. Instantiate two arbitrary boxes and then print their volume and surface area.
7.Create a class MathOperation that has four static methods. add() method that takes two integer numbers as
parameter and returns the sum of the numbers. subtract() method that takes two integer numbers as parameter and
returns the difference of the numbers. multiply() method that takes two integer numbers as parameter and returns the
product. power() method that takes two integer numbers as parameter and returns the power of first number to
second number. Create another class Demo (main class) that takes the two numbers from the user and calls all four
methods of MathOperation class by providing entered numbers and prints the return values of every method.